Added "hello world" go file
Creates an image based on randomly selected colors (not yet unique). The point of this application is to generate a list of all possible true colors (24bit colors) and display an image using all of the unique colors only once... not more.... not less.... once. Uniqueness will be handled next and then more algorithms and possibly a front end. This is going to be a new side project?